+++ CHANGELOG.rst
@@ -5,9 +5,9 @@
 * `#627 <https://github.com/xitrum-framework/xitrum/issues/627>`_
   Allow spaces and non-ASCII characters in static file names
 * `#629 <https://github.com/xitrum-framework/xitrum/issues/629>`_
-  Remove tf, tcf, tnf, and tcnf in I18n
+  Remove tf, tcf, tnf, and tcnf in I18n; replace them with added t, tc, tn, and tcn variants
 * `#630 <https://github.com/xitrum-framework/xitrum/issues/630>`_
-  Replace language (String) in I18n with locale (java.util.Locale)
+  Add I18n#locale (java.util.Locale)
 * `#628 <https://github.com/xitrum-framework/xitrum/issues/628>`_
   Update Netty from 4.1.3 to 4.1.4